requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=21.0.0
scipy>=1.10.0
massive>=0.1.0
websockets>=12.0
requests>=2.31.0
asyncio
plotly>=5.0.0
joblib>=1.3.0
pytest>=7.4.0
python-dotenv>=1.0.0
signalrcore>=0.9.5
httpx>=0.25.0
aiohttp>=3.9.0
eth-account>=0.10.0
